Move market and roll commands to new application
